+++ controllers/home.js
@@ -1,0 +1,7 @@
+const home = (req,res) =>{
+    res.render('start');
+}
+
+module.exports = {
+    home:home,
+}
