+++ routes/home.js
@@ -4,9 +4,6 @@ const home = require('../controllers/home');
 
 
 //Hier zet ik alle routes en zo zet ik ze weer naar een controller
-router.get('/', (req,res,) => {
-    res.render('start');
-
-});
+router.get('/', home.home)
 
 module.exports = router;
