+++ routes/index.js
@@ -5,13 +5,14 @@ const userRoutes = require('./users');
 const loginRoutes = require('./login');
 const registerRoutes = require('./register');
 const goalRoutes = require('./goals');
-const loggedUser = require('../middleware/isLoggedIn')
+const loggedUser = require('../middleware/authorization')
 
 
-router.use('/', homeRoutes);
+router.use('/welcome',loggedUser.guest, homeRoutes);
+
 router.use('/users', loggedUser.loggeduser, userRoutes);
 router.use('/register', registerRoutes);
-router.use('/login', loggedUser.loggeduser, loginRoutes);
+router.use('/login', loginRoutes);
 
 
 //page not found
